requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.5
openai==1.7.2

# Data fetching and API clients
requests==2.31.0
pycoingecko==3.1.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Environment management
python-dotenv==1.0.0

# CLI enhancements
rich==13.7.0

# Utilities
python-dateutil==2.8.2

